OLD APPS/requirements.txt
notion-client
feedparser
python-dotenv
requests 